Move WindowingInternals to runners/core-java
